Add metersPerUnit to RenderMan importer
This matches the hydra importer's operation. The hydra importer queries the scene for the scene units. RenderMan is unitless unless set in an IES profile. USD defaults to 1 cm as the unit which makes metersPerUnit = 0.01. When setting to that value in the importer, there was little difference in the lighting, but the shadow terminators became sharper. Leaving the value sets to 1 better match prman for now.