Add standard RenderMan OSL Patterns
Lifted straight from the RenderMan install, these OSL shaders mostly work straight out of the box with Cycles. Early testing indicates that the Mapping patterns (PxrNormalMap, PxrTexture, ...) do not work with Cycles. A preliminary guess is that it requires UV's of which there is no method to pass them into the pattern.