Fix bug with FaceVarying normals
This only would have ever worked for triangles. For a polygon with vertices > 3, the code would overwrite the normal for each generated triangle, and leave all triangles beyond the number of original polygons without any normals. First became apparent in the dishes in San Miguel with either the bottom being black, or the plate only being half lit.