Switch face-varying normals to cycles vertex
The default behavior (from the Hydra code) is to average the normals and create a face normal. However, if the face has a uv map, the tangent creation code expects vertex normals. So, switch the conversion to average the normals at the vertices until the tangent code is modified, or cycles supports face varying.