Honor the "shade int[N] faceset" attribute
When there is a single mesh with multiple materials attached to it, e.g., the silverware in the San Miguel scene, RenderMan creates a duplicate of the object for each of the materials and puts an attribute on it, "shade", that effectively masks pieces of geometry from the renderer. While Cycles appears to have a per-prim mask, I can't actually determine how (and if) it's used. So, this commit subsets the entity and creates a new one based on the values in the "shade" attribute.