Cycles normal maps don't work with PxrBlender
It should be a pass through, but PxrBlenderPrincipledInputs interferes with the downstream utilization of the normal map (normals are 0). So now we look if PxrBlender is hooked to a normal map, and the downstream OSL (lama) node has a connection from a "normal*" input to PxrBlender, we directly connect the normal map to the OSL node and bypass PxrBlender.

Only currently works with converted Lama networks.